requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.0
Werkzeug==2.3.7
pytest==7.4.2
pytest-cov==4.1.0
python-dotenv==1.0.0
redis==5.0.1